pyproject.toml
[tool.poetry]
name = "metis"
version = "1.0.0"
description = ""
authors = ["Eneko Illarramendi <eneko@illarra.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "~3.12.0"
django = {extras = ["bcrypt"], version = "*"}
djangorestframework = "*"
django-allauth = "==0.61.1"
django-compressor = "*"
django-cors-headers = "*"
django-countries = "*"
django-ical = "*"
django-libsass = "*"
django-modeltranslation = "*"
django-storages = {extras = ["s3"], version = "*"}
django-vite = "*"
drf-extensions = "*"
drf-writable-nested = "*"
gevent = "*"
gunicorn = "*"
hiredis = "*"
holidays = "*"
huey = "*"
inertia-django = "*"
markdown = "*"
mysqlclient = "*"
nh3 = "*"
openpyxl = "*"
pandas = "*"
phonenumbers = "*"
pydantic = "*"
pyyaml = "*"
redis = "*"
requests = "*"
scipy = "*"
sentry-sdk = {extras = ["django"], version = "*"}
tzdata = "*"
uritemplate = "*"
whitenoise = "*"
xhtml2pdf = "*"

[tool.poetry.group.dev.dependencies]
arrow = "*"
django-debug-toolbar = "*"
factory-boy = "*"
phonenumbers = "*"
pytest = "*"
pytest-cov = "*"
pytest-django = "*"
pytest-xdist = "*"
responses = "*"
ruff = "*"
sqlalchemy = "*"

[tool.coverage.run]
omit = [
  "*/management/**",
  "*/migrations/**",
  "*/settings/**",
  "*/tests/**",
  "*apps.py",
  "*urls.py",
  "*wsgi.py",
  "manage.py",
]

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "raise NotImplementedError",
  "def __repr__",
  "def __str__(self)",
  "if TYPE_CHECKING:"
]
show_missing = true

[tool.pytest.ini_options]
addopts = "--cov=metis --cov-report=html"
markers = [
  "api: mark a test as an `api` test",
  "site: mark a test as an `site` test",
  "unit: mark a test as an `unit` test",
]
DJANGO_SETTINGS_MODULE = "metis.settings.test"

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "UP", "B", "SIM", "I", "D", "DJ"]
ignore = ["SIM105", "D100", "D104", "D105", "D107", "D417", "D203", "D213"]

[tool.ruff.lint.isort]
lines-after-imports = 2
